Time: 199 ms (33.33%) | Memory: 111.5 MB (54.49%) - LeetSync

+++ 295-find-median-from-data-stream/find-median-from-data-stream.java
@@ -24,7 +24,7 @@ class MedianFinder {
 
         if(minHeap.size() == maxHeap.size())
             return ((maxHeap.peek()+minHeap.peek())/(double)2);
-        return (double)maxHeap.peek();
+        return maxHeap.peek();
     }
 }
